// src/shared/models/shared-brand.model.ts
import { z } from 'zod'

export const BrandSchema = z.object({
	id: z.number(),
	name: z.string().max(500),
	logo: z.string().url().max(1000),

	createdById: z.number().nullable(),
	updatedById: z.number().nullable(),
	deletedById: z.number().nullable(),
	deletedAt: z.date().nullable(),
	createdAt: z.date(),
	updatedAt: z.date(),
})

export type BrandType = z.infer<typeof BrandSchema>