// src/shared/models/shared-category.model.ts
import { z } from 'zod'

export const CategorySchema = z.object({
	id: z.number(),
	parentCategoryId: z.number().nullable(),
	name: z.string(),
	logo: z.string().nullable(),

	createdById: z.number().nullable(),
	updatedById: z.number().nullable(),
	deletedById: z.number().nullable(),
	deletedAt: z.date().nullable(),
	createdAt: z.date(),
	updatedAt: z.date(),
})